refactor: controllers in dedicated classes
move the control and feedback logic into an abstract class with pure
virtual functions, and add the class PodController that implements the
abstract part of Controller to run on PodController
